requirements.txt
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
pillow
psycopg
python-decouple